Omni directional shadow map
Added Omni Directional Shadows

+++ DirectionalLight.h
@@ -6,7 +6,7 @@ class DirectionalLight :
 {
 public:
 	DirectionalLight();
-	DirectionalLight(GLuint shadowWidth, GLuint shadowHeight,
+	DirectionalLight(GLfloat shadowWidth, GLfloat shadowHeight,
 		GLfloat red, GLfloat green, GLfloat blue,
 		GLfloat aIntensity, GLfloat dIntensity,
 		GLfloat xDir, GLfloat yDir, GLfloat zDir);
